+++ template-parts/home/hp-sponsor-section.php
@@ -14,6 +14,7 @@ if ( 'true' === $section_enabled ) {
 	);
 	$num_items = $query->post_count;
 	if ( $num_items > 0 ) {
+		$items_per_row = dli_get_option( 'num_row_sponsor', 'sponsors' );
 ?>
 
 
@@ -26,8 +27,16 @@ if ( 'true' === $section_enabled ) {
 						$image_metadata = dli_get_image_metadata( $post, 'full', '/assets/img/yourimage.png' );
 						$post_id        = $post->ID;
 						$external_link  = dli_get_field( 'link_esterno', $post_id );
+						if ( $items_per_row === '4' ){
 					?>
-					<div class="col-6 col-lg-2">
+					 <div class="col-4 col-lg-3">
+					<?php
+						} else {
+					?>
+						<div class="col-6 col-lg-2">
+					<?php
+						}
+					?>
 						<div class="it-grid-item-wrapper">
 							<a href="<?php echo esc_attr( $external_link ); ?>" target="_blank">
 								<figure class="figure img-full w-100">