+++ CHANGELOG.md
@@ -21,7 +21,7 @@ TAGS: Aggiunto (Added), Modificato (Changed), Deprecato (Deprecated), Rimosso (R
 - Corretto layout del dettaglio di una Risorsa Tecnica.
 ## Modificato
 - Estesa visualizzazione senza foto delle persone anche a progetti e attività di ricerca (hide_person_icon).
-- Parametrizzata gestione della sezione Sponsor.
+- Parametrizzata gestione della sezione degli Sponsor in Home Page.
 ### Aggiunto
 - Aggiunto contenuto: impatto.